hasBody middleware bug and update
This conditional used to only apply to POST requests, but after I
started using other HTTP methods in routes/api.js, this condition
needed an update to check those too.

The problem:
url.split("/")[1] returns "api" if the url is "/api/auth/connect",
when it's supposed to return "auth". This continued the check for
a request body when it shouldn't have.

The solution:
url.split("/")[2] !== "auth"  OR  url.indexOf("auth") === -1
The first would work, but the second is a bit simpler.

As a reminder, "auth" requests are an exception to hasBody because
both of the auth POST requests don't have bodies.