Update page handlers to stop gathering data
Page handlers used to grab data from the database and render the page
with that data directly. Now, pages with no preprocessed data are sent
to the user, and data is loaded into these pages later through the API.

This will speed up display of pages to end users; it was quite slow
before.