Update view JavaScript to display API data
View JavaScript now sends requests to get API data for each respective
view. Once the data is received, it is displayed accordingly through
vanilla JavaScript, and the appropriate loading spinner is hidden from
view.

For the entries & results pages, the sidebar now correctly highlights
which contest is "selected" according to the URL.